refactor(di): init app using plugin-baby
